third_party/blink/web_tests/transforms/3d/hit-testing/coplanar-with-camera-expected.txt

The text on this element should be selectable. Hovering on this element should cause a highlight.
Element at 98, 200 has id "background": PASS
Element at 302, 200 has id "background": PASS
Element at 200, 98 has id "background": PASS
Element at 200, 302 has id "background": PASS
Element at 101, 200 has id "layer": PASS
Element at 299, 200 has id "layer": PASS
Element at 200, 101 has id "layer": PASS
Element at 200, 299 has id "layer": PASS
